feat: improve recommendation system by processing the full lyrics to get the whole context of a song
