+++ app/lib/generateSeedPlaylist.ts
@@ -122,8 +122,8 @@ export async function generateSeedPlaylist(
 				// p-limit concurrency of 15
 				const limit = pLimit(15);
 
-				// 0.75 is a good starting point, but you can tune this
-				const THRESHOLD = 0.75;
+				// 0.8 is a good starting point, but you can tune this
+				const THRESHOLD = 0.8;
 
 				// we want to avoid low quality recommendations so we check each song against all seeds(selected songs) put that into an array and sort by the number of seeds that match the treshold then return the top 100
 				// we also want to avoid songs that are too far in similarity to the seeds so we set a cutoff of 0.2 below the threshold to give extra grace
